Refactor - folder structure
