+++ .gitignore
@@ -13,6 +13,7 @@ develop-eggs/
 downloads/
 eggs/
 .eggs/
+*.egg-info
 lib/
 lib64/
 parts/
