+++ AutoApiTestRunner.egg-info/SOURCES.txt
@@ -1,5 +1,7 @@
 README.md
 setup.py
+AutoApiTestRunner/__init__.py
+AutoApiTestRunner/config.py
 AutoApiTestRunner.egg-info/PKG-INFO
 AutoApiTestRunner.egg-info/SOURCES.txt
 AutoApiTestRunner.egg-info/dependency_links.txt
